requirements.txt
fuzzywuzzy
gspread
streamlit>=1.20.0
streamlit-aggrid
extra_streamlit_components
matplotlib
google-cloud-bigquery[pandas]==2.34.3
google-auth
db-dtypes